fix(clash): shrink step.ts and duplicates.ts under the module-size gate
The merge from upstream/main plus this PR's own additions pushed both
files past their recorded module-size budgets (step.ts 446 vs 414,
duplicates.ts 436 vs 433), failing check-module-size.mjs.

- Split buildStepExclusions() out of step.ts into a new sibling
  step-exclusions.ts, re-exported from step.ts so the
  @ifc-lite/clash/step public surface is unchanged. step.ts drops to
  399 lines; its allowlist row is deleted per the gate's own rule.
- Tightened the sweep-key comment in duplicates.ts (no information
  dropped) to bring it back to exactly its 433-line budget.
