+++ packages/clash/src/adapters/step.ts
@@ -17,10 +17,11 @@
 import { type IfcDataStore } from '@ifc-lite/parser';
 import { EntityNode } from '@ifc-lite/query';
 import type { MeshData } from '@ifc-lite/geometry';
-import { makeExclusionSet, qualifiedKey } from '../exclude.js';
+import { makeExclusionSet } from '../exclude.js';
 import { fromPositions, NonFiniteAxisError } from '../math/aabb.js';
 import type { ClashElement, ExclusionSet, Mat4 } from '../types.js';
 import { isNonClashableTag, mergeMeshes } from './shared.js';
+import { buildStepExclusions } from './step-exclusions.js';
 
 /** Minimal federation contract — pass an `@ifc-lite/renderer` `FederationRegistry`. */
 export interface FederationLike {
@@ -392,55 +393,7 @@ export function elementsFromStep(options: StepAdapterOptions): StepAdapterResult
   return { elements, exclusions };
 }
 
-/**
- * Pair-exclusions from IFC relationships. Only relationship getters
- * (`voids`/`filledBy`/`decomposedBy`/`decomposes`) are used here; these read
- * the relationship graph and never call `extractEntityAttributesOnDemand`, so
- * the per-element loop stays off the AGENTS.md hot-loop anti-pattern:
- * - host vs the filler of its opening (wall vs door/window)
- * - element vs its own (meshed) opening
- * - members of the same `IfcRelAggregates` assembly
- */
-export function buildStepExclusions(
-  store: IfcDataStore,
-  byExpressId: Map<number, ClashElement[]>,
-): ExclusionSet {
-  const pairs: Array<[string, string]> = [];
-
-  for (const [expressId, elementsAtId] of byExpressId) {
-    const node = new EntityNode(store, expressId);
-
-    // A relationship is stated between EXPRESS ids, not occurrences: fan it
-    // out across every occurrence bucketed at each side (usually one element
-    // each; more than one only for a GPU-instanced expressId), so a host's
-    // void/assembly exclusions cover every physical placement of the
-    // filler/sibling, not just whichever occurrence happened to be built last.
-    const pairAll = (otherId: number): void => {
-      const others = byExpressId.get(otherId);
-      if (!others) return;
-      for (const a of elementsAtId) {
-        const ek = qualifiedKey(a.model, a.key);
-        for (const b of others) {
-          pairs.push([ek, qualifiedKey(b.model, b.key)]);
-        }
-      }
-    };
-
-    for (const opening of node.voids()) {
-      pairAll(opening.expressId);
-      for (const filler of opening.filledBy()) {
-        pairAll(filler.expressId);
-      }
-    }
-
-    const parent = node.decomposedBy();
-    if (parent) {
-      for (const sibling of parent.decomposes()) {
-        if (sibling.expressId === expressId) continue;
-        pairAll(sibling.expressId);
-      }
-    }
-  }
-
-  return makeExclusionSet(pairs);
-}
+// Pair-exclusions from IFC relationships (voids/filledBy/decomposedBy/decomposes)
+// live in `./step-exclusions.js`, split out purely to keep this file under its
+// module-size budget; re-exported here so `@ifc-lite/clash/step` is unchanged.
+export { buildStepExclusions } from './step-exclusions.js';
